Add files via upload
This project aims to explore the Walmart Sales data to understand top performing branches and products, sales trend of of different products, customer behaviour. The aims is to study how sales strategies can be improved and optimized. The major aim of thie project is to gain insight into the sales data of Walmart to understand the different factors that affect sales of the different branches